Conflicts:
	CMakeLists.txt
	src/servers/Server_Common/Network/PacketDef/Ipcs.h
	src/servers/Server_Zone/Actor/PlayerQuest.cpp
	src/servers/Server_Zone/Actor/PlayerSql.cpp

+++ sql/infolinkshell.sql
@@ -23,7 +23,7 @@ DROP TABLE IF EXISTS `infolinkshell`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `infolinkshell` (
-  `LinkshellId` int(20) NOT NULL AUTO_INCREMENT,
+  `LinkshellId` bigint(20) NOT NULL AUTO_INCREMENT,
   `MasterCharacterId` int(20) DEFAULT NULL,
   `CharacterIdList` blob,
   `LinkshellName` varchar(32) DEFAULT NULL,
