+++ src/servers/Server_Zone/Linkshell/LinkshellMgr.h
@@ -12,7 +12,7 @@ typedef boost::shared_ptr< Linkshell > LinkshellPtr;
 class LinkshellMgr
 {
 private:
-   std::map< uint32_t, LinkshellPtr > m_linkshellIdMap;
+   std::map< uint64_t, LinkshellPtr > m_linkshellIdMap;
    std::map< std::string, LinkshellPtr > m_linkshellNameMap;
 
    LinkshellPtr getLinkshellByName( const std::string& name );
